Specify dtypes for NumPy arrays
The default dtype in NumPy is float64, or a 64-bit floating point number. These are very large and not that fast in calculations.

In this commit a dtype is defined for every array, with a fitting type and precision. Often float32 is used, which is the fastest in calculations (fully hardware accelerated) and has only half the memory footprint of float64. Sometimes an even faster int or uint (for counts) or bool (for masks) is used.