JOSS paper: Consider adding hyperlink to my GitHub
Awesome to hear you have submitted a JOSS paper! Thanks for referring to me explicitly in the acknowledgement, that's appreciated!

I have a small request: Would you mind adding a hyperlink to my GitHub profile?

If not that's also okay, this would just make me a little bit more findable.